Initial Commit of Project
Uploading my own copy of the 'CyberDino' game project I worked on helping to script in college. While the digital media students created the dinosaurs' animations, rigid bodies and textures, the computer science students (me included) gave them a soul. The game was quickly established as an award-winning piece of student work. I was glad to be a part of it.